requirements.txt
openai>=1.50.2
python-dotenv
beautifulsoup4
requests
schedule
yagmail
fpdf
